fix(gate-3b): extract test modules + update P-03 to 14 tools (#608)
- graph_read_neighbors.rs: extract #[cfg(test)] to graph_read_neighbors_tests.rs (356 lines, was 506)
- graph_queries.rs: extract test module to graph_queries_tests.rs and neighbor SQL helpers to graph_queries_neighbors.rs (450 lines, was 1045)
- test_protocol.py: rename test_list_tools_returns_thirteen → test_list_tools_returns_fourteen, add context_graph to expected list
